+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- add -vv flag for extra verbose output
+###Changed
+- -last_info option: skip model transfer if last updated model buffer and the current model buffer overlap exactly
+
 ## [0.11] - 26-04-2023
 ### Added
 - add -b or --binary option for read and write files in binary format
@@ -8,7 +14,7 @@ All notable changes to this project will be documented in this file.
 - add tests for the compression entity
 - add test for cmp_rdcu_cfg.c
 - add sparc inttypes.h definitions
-###Changed
+### Changed
 - refactor function and constants names
 - refactor configure check functions
 - move the rdcu setup functions to cmp_rdcu_cfg.c
@@ -21,14 +27,13 @@ All notable changes to this project will be documented in this file.
 - fixed several bug when using the last_info option
 - fix a bug in the calculation of the adaptive compression sizes
 
-
 ## [0.09] - 30-09-2022
 ### Added
 - decompression/compression for non-imagette data
 - functions to create and configure a compression configuration
 - add max_used_bits feature
 - add max used bit version field to the compression entity
-###Changed
+### Changed
 - Change the build system form make to meson
 - Change DEFAULT_CFG_MODEL and DEFAULT_CFG_DIFF to CMP_DIF_XXX constants
 ### Fixed